🐛 fix: a fixed-width integer settles by its type, a float by where it is stored
C# wraps a byte past 255 and throws in a checked context; JavaScript keeps counting in a
double. The compiler now reads the RESULT type of every + - * << (and ++, -- and the
compound forms) and settles the value where C# would (IntegerWidth): byte, sbyte, short,
ushort and uint always wrap (& 0xFF, << 16 >> 16, >>> 0, Math.imul for a 32-bit product —
the FNV step on a uint is the step it is in .NET); int and long wrap only under an
explicit `unchecked`, because every plain `i + 1` in a UI would otherwise carry a `| 0`
for an overflow that is a bug anywhere else; a `checked` context throws through
$eq.num.checked. Whether the context is checked comes from the bound tree —
IOperation.IsChecked via SemanticHelper.GetOperation, the first real use of IOperation —
so the project-wide setting, checked blocks and checked(...) expressions all count
(ArithmeticContext).

A float rounds to single precision at the seams ECMA-335 guarantees it (FloatStore): a
declaration or assignment, a float operand of a comparison, and ToString, where
$eq.num.single prints the shortest decimal that reads back as the same single — 0.1f +
0.2f is "0.3" on both sides. Rounding every arithmetic instead wrapped 16 shared
components in Math.fround; rounding at the store touches 7, all `let x = Math.fround(…)`.

char++ steps the code unit and `sum += ch` adds it; an enum in arithmetic or CompareTo
computes on the value behind the member name (the Cast strategy's maps, now shared).

The conversion instrument grows 66 → 76 cases (unchecked wrap, checked throw, the FNV
step, float stores and comparisons) and its gap baseline shrinks 11 → 2: the two are the
documented limit — int.MaxValue + 1 and long.MaxValue + 1 with no `unchecked`.

+++ src/eQuantic.UI.Runtime/src/shared/components/Mermaid.ts
@@ -55,17 +55,17 @@ export class Mermaid extends StatelessComponent {
     }
 
     static placeArrowhead(arrow: MermaidArrowhead, ink: ColorToken) {
-        let half = (Mermaid.headSize - 1) / 2;
-        let top = arrow.y - Mermaid.headSize;
-        let start = arrow.x - half;
+        let half = Math.fround((Mermaid.headSize - 1) / 2);
+        let top = Math.fround(arrow.y - Mermaid.headSize);
+        let start = Math.fround(arrow.x - half);
         if (arrow.direction === 1) {
-            top = arrow.y - half;
-            start = arrow.x - Mermaid.headSize;
+            top = Math.fround(arrow.y - half);
+            start = Math.fround(arrow.x - Mermaid.headSize);
         } else if (arrow.direction === 2) {
             top = arrow.y;
-            start = arrow.x - half;
+            start = Math.fround(arrow.x - half);
         } else if (arrow.direction === 3) {
-            top = arrow.y - half;
+            top = Math.fround(arrow.y - half);
             start = arrow.x;
         }
         return new Positioned(new Vector(Mermaid.headGlyph(arrow.direction), Mermaid.headSize, ink), top, null, null, start);
